Fix expected type of id field in ProductGalleryImageUploadResult and update obsolete tests

+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/product/request/ProductGalleryImageDeleteRequest.kt
@@ -5,7 +5,7 @@ import com.ecwid.apiclient.v3.impl.RequestInfo
 
 data class ProductGalleryImageDeleteRequest(
 	val productId: Int = 0,
-	val fileId: Int = 0
+	val fileId: Long = 0
 ) : ApiRequest {
 	override fun toRequestInfo() = RequestInfo.createDeleteRequest(
 		pathSegments = listOf(
